+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -9,7 +9,7 @@ CHANGELOG
  * added `Symfony\Component\Form\FormErrorIterator::findByCodes()`
  * added `getTypedExtensions`, `getTypes`, and `getTypeGuessers` to `Symfony\Component\Form\Test\FormIntegrationTestCase`
  * added `FormPass`
- * added `compare` option to `RepeatedType`
+ * added `comparator` option to `RepeatedType`
 
 3.2.0
 -----